Add files via upload
These three files are part of an end-to-end system of splitting, labeling, pre-processing, training and testing a machine learning CNN classifier. I tried to code these scripts in such a way that my colleagues at NASA Goddard could continue to use my code after I left. As a result, most do not work unless you have an appropriate data set to use with the files. The script SplitnClass_tool_GUI_OFF.py is a general user interface coded in python that was meant to aid the scientists I was working with in labeling the data set. The code allows you to select your source folder then loops through each image in the folder, slicing each image and displaying the parent image and the current slice on the screen. The user may then classify the current slice by pressing one of the buttons on the menu at which time that slice will vanish and the next slice will be displayed. This continues until the user quits the program or the folder runs out of images to slice. The code also creates a simple record text file that is stored in the source folder upon program execution. The program records the filenames of all images processed and skips any files with these filenames if the program is quit and rerun with the same source folder. Train_Test_Creator is a simple program that creates training and testing data sets using the files in the folder selected by the user. It creates two new sub-directories in the directory then populates these with the appropriate split of data corresponding with either training or testing respectively. Since the python code in the machine learning script associates the class label with the name of the folder that the data is retrieved from, the existing label folders in the original directory are replicated in the training and testing directories. Finally, TPR_ML_v1.0.4.py is the script containing the actual machine learning code. Using a python library called Keras, I import an existing model with the intent of using transfer learning to improve accuracy and reduce training time of the final model (this project had a rather small data set so it was hoped that transfer learning would offset the associated shortcomings). The imported model is augmented to provide a classifier for our data set by adding a new dense layer to the end of the network. Finally, all the layers are set to be trainable which increases training time but is intended to provide better final accuracy. The script is finished with testing and training commands and code is included to evaluate the performance of the classifier.