Add files via upload
This is a matlab script for implementing a Fully Convolutional Neural Network (FCNN) for the Textural Pattern Recognition project. This script was created in addition to the python scripts so that there would be an option for segmentation of the input images instead of purely classification. This script also uses transfer learning and takes advantage of a built in model available in matlab. It also uses built in functions to import, pre-process train and evaluate the machine learning model all in the same script.